+++ strato/function/template.php
@@ -13,6 +13,11 @@ namespace BurningMoth\Stratus;
  */
 function template( $___template_path, array $___template_variables = array() ) {
 
+	global $tratus;
+
+	// load from extended structure ...
+	if ( preg_match('/^[\w-]+$/', $___template_path) ) $___template_path = sprintf('strato/template/%s.php', $___template_path);
+
 	// dump vars ...
 	if ( $___template_variables ) extract($___template_variables);