+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # STRATUS CHANGELOG #
 
+### 2019 May 25 / Version 2.1
+- Added: classes `cURL`, `RandomString`
+- Updated: classes `ArrayRef`
+- Added: functions `constant`, `cookie_domain`, `fwdslashes`, `get_arg`, `head_arg`, `server_arg`, `post_arg`, `path_arg`, `leadingslash`, `trailingslash`, `mail`, `random_int`, `random_str`, `session_exists`, `url`
+- Fixed: `Hook.exec()` breaking on undefined value.
+- Added: `String.substrCount()`, `String.htmlEncode()`, `String.htmlDecode()`
+
 ### 2018 Nov 25 / Version 2.0
 - Removed: all WordPress-specific functionality.
 - Added: native action and filter hooks.